requirements.txt
agate==1.9.1
altair==6.0.0
annotated-types==0.7.0
asn1crypto==1.5.1
attrs==26.1.0
babel==2.18.0
beautifulsoup4==4.14.3
blinker==1.9.0
boto3==1.42.75
botocore==1.42.75
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
confluent-kafka==2.13.2
curl_cffi==0.13.0
daff==1.4.2
dbt-adapters==1.22.10
dbt-common==1.37.3
dbt-core==1.11.7
dbt-extractor==0.6.0
dbt-postgres==1.10.0
dbt-protos==1.0.443
dbt-redshift==1.10.1
dbt-semantic-interfaces==0.9.0
deepdiff==8.6.2
frozendict==2.4.7
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
importlib_metadata==8.9.0
isodate==0.7.2
Jinja2==3.1.6
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
leather==0.4.1
lxml==5.4.0
MarkupSafe==3.0.3
mashumaro==3.14
more-itertools==10.8.0
msgpack==1.1.2
multitasking==0.0.12
narwhals==2.18.1
networkx==3.6.1
numpy==2.4.3
orderly-set==5.5.0
packaging==26.0
pandas==2.3.3
parsedatetime==2.6
pathspec==0.12.1
peewee==4.0.2
pillow==12.1.1
platformdirs==4.9.4
plotly==6.6.0
protobuf==6.33.6
psycopg2-binary==2.9.11
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2026.1.post1
PyYAML==6.0.3
redshift-connector==2.1.12
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
s3transfer==0.16.0
scramp==1.4.8
six==1.17.0
smmap==5.0.3
snowplow-tracker==1.1.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
sqlparse==0.5.4
streamlit==1.55.0
tenacity==9.1.4
text-unidecode==1.3
toml==0.10.2
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==16.0
yfinance==1.2.0
zipp==3.23.0
